Add a prompt mode to allow manual selection of results.
Passing the `--prompt` flag will ask the user at the end of each
test whether that run should be considered a regression or not.
They can also choose to retry the test before deciding.

This is particularly useful when using `cargo_args` to run
arbitrary cargo commands like benchmarks which may not include
all of the pass/fail information in their exit code and which
may require human interpretation.