note that specified commits should be from the master branch
While trying to get cbr working I ran into the ominous
"Expected author Brian Anderson to be bors for beb9a0dfc52ebda4f8db4e5d439e08e4f3a43a39"
error.

The problem seems to be that all commits are expected to be found in the master branch, so simply feeding in a git tag commit or
something of a beta/stable branch does not work.

Tweak error message to reflect that.

cc #27