tests: Fix tests if the HOST triple is too long
The HOST triple defaults to the current host and gets printed in the help output. The CLI tests
currently match it with `[..]`. However, if the triple is too long clap might introduce a line
break, which `[..]` can't handle.
Use `...` instead to match arbitrary lines at the position. Not elegant, but these are just smoke
tests to ensure there's no major probelm. They are not meant to ensure 100% stable output.

Fixes: https://github.com/rust-lang/cargo-bisect-rustc/issues/226