Stop stripping and LTO
*Stop stripping*:
Stripping symbols here is bad, because it makes the binary completely opaque and undebbuggable, should something ever go wrong. The binary size benefit is very minor, this program is supposed to run on desktop computers with tons of disk space.
The tradeoff is bad.

*Stop LTO*:
The same binary size non-concerns as stripping apply. Also, speed basically doesn't matter here.
Meanwhile, we want this program to be `cargo install`-able, so compile times matter.
Again, the tradeoff is bad.