Don't allow Bounds::Commits to contain non-hashes.
There was only ever one case of it containing a non-hash, and that was
when `--start` was a hash, and `--end` was not specified. In that case,
it was "origin/master", and then converted to a hash later on. To
simplify things, this converts `origin/master` to master's actual commit
early when the `Bounds` is created.